level 0
The goal of this level is for you to log into the game using SSH. The host to which you need to connect is bandit.labs.overthewire.org, on port 2220. The username is bandit0 and the password is bandit0. Once logged in, go to the Level 1 page to find out how to beat Level 1.

bandit 문제를 풀기에 앞서, terminal 'putty'을 다운로드하였습니다. 
다운로드 링크 https://www.chiark.greenend.org.uk/~sgtatham/putty/latest.html

Host Name (or IP address): bandit.labs.overthewire.org
Port: 2220

login as(ID): bandit0
password: bandit0